Created empty packet handlers for now. Tried adding a pch but that didnt work. damn c files, ill try to get rid of aes later. now processing packets immediately instead of storing in a queue, for now... ill figure that out eventually

+++ src/net/packets/PacketHandler.h
@@ -7,7 +7,7 @@
 
 namespace net
 {
-    typedef std::unordered_map<byte, std::function<void(Packet&)>> PacketHandlers;
+    typedef std::unordered_map<uint16_t, std::function<void(Packet&)>> PacketHandlers;
 
     class PacketHandler
     {
